+++ src/sketch_topo.cpp
@@ -1,0 +1,759 @@
+#include "sketch_topo.h"
+
+#include <BRepBuilderAPI_MakeEdge.hxx>
+#include <BRepBuilderAPI_MakeFace.hxx>
+#include <BRepBuilderAPI_MakeWire.hxx>
+#include <BRepTools.hxx>
+#include <GC_MakeArcOfCircle.hxx>
+#include <Graphic3d_AspectFillArea3d.hxx>
+#include <Precision.hxx>
+#include <Quantity_Color.hxx>
+#include <TopoDS.hxx>
+#include <TopoDS_Wire.hxx>
+#include <algorithm>
+#include <functional>
+#include <limits>
+#include <map>
+#include <unordered_map>
+#include <unordered_set>
+
+#include <glm/glm.hpp>
+
+#include "occt_view.h"
+#include "sketch.h"
+#include "sketch_delta.h"
+#include "utl.h"
+#include "utl_geom.h"
+
+using namespace glm;
+
+struct Sketch_topo::Face_edge
+{
+  const Sketch::Edge& edge;
+  bool                reversed;
+
+  size_t start_nd_idx() const;
+  size_t end_nd_idx() const;
+};
+
+Sketch_topo::Sketch_topo(Sketch& sketch)
+    : m_sketch(sketch)
+{
+}
+
+void Sketch_topo::remove_displayed_faces()
+{
+  for (AIS_Shape_ptr& f : m_faces)
+    m_sketch.m_ctx.Remove(f, false);
+
+  m_faces.clear();
+  m_dim_classifier_faces.clear();
+}
+
+double Sketch_topo::plane_pick_snap_radius_world() const
+{
+  const double px = Sketch_nodes::get_snap_dist();
+  if (m_sketch.m_view.is_headless())
+    return std::max(px, Precision::Confusion() * 1e9);
+
+  const gp_Pnt2d          ref2(0.0, 0.0);
+  const ScreenCoords      s0   = m_sketch.m_view.get_screen_coords(to_3d(m_sketch.m_pln, ref2));
+  const dvec2             base = s0.unsafe_get();
+  std::optional<gp_Pnt2d> p1   = m_sketch.m_view.pt_on_plane(ScreenCoords(dvec2(base.x + px, base.y)), m_sketch.m_pln);
+  if (!p1)
+    return std::max(px, Precision::Confusion() * 1e9);
+
+  return ref2.Distance(*p1);
+}
+
+void Sketch_topo::snap_placed_node_to_closest_linear_edge_interior_(size_t node_idx)
+{
+  EZY_ASSERT(node_idx < m_sketch.m_nodes.size());
+
+  const gp_Pnt2d p(m_sketch.m_nodes[node_idx].X(), m_sketch.m_nodes[node_idx].Y());
+  const double   max_d = plane_pick_snap_radius_world();
+
+  double                  best_err = std::numeric_limits<double>::infinity();
+  std::optional<gp_Pnt2d> best_proj;
+
+  for (const auto& e : m_sketch.m_edges)
+  {
+    if (e.circle_arc)
+      continue;
+
+    if (e.node_idx_arc.has_value())
+      continue;
+
+    if (!e.node_idx_b.has_value())
+      continue;
+
+    const size_t    a  = e.node_idx_a;
+    const size_t    b  = *e.node_idx_b;
+    const gp_Pnt2d& pa = m_sketch.m_nodes[a];
+    const gp_Pnt2d& pb = m_sketch.m_nodes[b];
+    if (node_idx == a || node_idx == b)
+      continue;
+
+    std::optional<gp_Pnt2d> proj = snap_foot_to_open_segment_interior_if_close(p, pa, pb, max_d);
+    if (!proj)
+      continue;
+
+    const double err = p.Distance(*proj);
+    if (err < best_err)
+    {
+      best_err  = err;
+      best_proj = proj;
+    }
+  }
+
+  if (best_proj)
+  {
+    Sketch_nodes::Node& n = m_sketch.m_nodes[node_idx];
+    n.SetX(best_proj->X());
+    n.SetY(best_proj->Y());
+    n.midpoint = false;
+  }
+}
+
+void Sketch_topo::split_linear_edges_at_node_if_interior(size_t node_idx)
+{
+  split_linear_edges_at_node_if_interior(node_idx, static_cast<Sketch_op_recorder*>(nullptr));
+}
+
+void Sketch_topo::split_linear_edges_at_node_if_interior(size_t node_idx, Sketch_op_recorder& rec)
+{
+  split_linear_edges_at_node_if_interior(node_idx, &rec);
+}
+
+void Sketch_topo::split_linear_edges_at_node_if_interior(size_t node_idx, Sketch_op_recorder* rec)
+{
+  EZY_ASSERT(node_idx < m_sketch.m_nodes.size());
+  snap_placed_node_to_closest_linear_edge_interior_(node_idx);
+  const gp_Pnt2d& p = m_sketch.m_nodes[node_idx];
+
+  bool progress = true;
+  while (progress)
+  {
+    progress = false;
+    for (auto itr = m_sketch.m_edges.begin(); itr != m_sketch.m_edges.end(); ++itr)
+    {
+      if (itr->circle_arc)
+        continue;
+
+      if (itr->node_idx_arc.has_value())
+        continue;
+
+      if (!itr->node_idx_b.has_value())
+        continue;
+
+      const size_t    a  = itr->node_idx_a;
+      const size_t    b  = *itr->node_idx_b;
+      const gp_Pnt2d& pa = m_sketch.m_nodes[a];
+      const gp_Pnt2d& pb = m_sketch.m_nodes[b];
+
+      if (node_idx == a || node_idx == b)
+        continue;
+
+      if (!point_on_open_segment_2d(p, pa, pb))
+        continue;
+
+      Sketch::Edge edge_a{a};
+      m_sketch.update_edge_end_pt_(edge_a, node_idx);
+      Sketch::Edge edge_b{node_idx};
+      m_sketch.update_edge_end_pt_(edge_b, b);
+
+      if (rec)
+        rec->note_prev_linear_edge(itr->node_idx_a, *itr->node_idx_b, itr->node_idx_mid, itr->name);
+
+      m_sketch.m_ctx.Remove(itr->shp, false);
+      m_sketch.m_edges.erase(itr);
+      m_sketch.m_nodes[node_idx].midpoint = false;
+      m_sketch.m_edges.push_back(std::move(edge_a));
+      m_sketch.m_edges.push_back(std::move(edge_b));
+
+      progress = true;
+      break;
+    }
+  }
+}
+
+// Function to extract faces from the planar graph
+void Sketch_topo::update_faces()
+{
+  m_sketch.m_nodes.finalize();
+
+  m_sketch.m_view.remove(m_faces);
+  m_faces.clear();
+  m_dim_classifier_faces.clear();
+
+  // Used to cleanup dangling nodes;
+  std::vector<bool> used_nodes(m_sketch.m_nodes.size());
+
+  // Build adjacency list
+  std::unordered_map<size_t, std::vector<std::pair<size_t, const Sketch::Edge*>>> adj_list;
+
+  for (const auto& edge : m_sketch.m_edges)
+  {
+    EZY_ASSERT(edge.node_idx_b.has_value());
+
+    size_t a = edge.node_idx_a;
+    size_t b = edge.node_idx_b.value();
+    adj_list[a].push_back({b, &edge});
+    adj_list[b].push_back({a, &edge});
+
+    // Keep track of used nodes.
+    used_nodes[a] = true;
+    used_nodes[b] = true;
+    if (edge.node_idx_mid.has_value())
+      used_nodes[*edge.node_idx_mid] = true;
+
+    if (edge.node_idx_arc.has_value())
+      used_nodes[*edge.node_idx_arc] = true;
+  }
+
+  if (m_sketch.m_operation_axis.has_value())
+  {
+    used_nodes[m_sketch.m_operation_axis->node_idx_a] = true;
+    if (m_sketch.m_operation_axis->node_idx_b.has_value())
+      used_nodes[*m_sketch.m_operation_axis->node_idx_b] = true;
+  }
+
+  // Remove dangling edges (edges with degree-1 endpoints) iteratively.
+  // These edges cannot form closed faces, so we exclude them from face detection.
+  std::unordered_set<const Sketch::Edge*> excluded_edges;
+
+  // Treat edges attached to the virtual mid-node of an arc as dangling for face detection.
+  // Circle arcs are represented using a virtual node in the middle of the arc (`node_idx_arc`).
+  // Any non-arc edge that uses this virtual node as an endpoint should not participate in faces.
+  {
+    std::unordered_set<size_t> arc_mid_nodes;
+    for (const auto& edge : m_sketch.m_edges)
+      if (edge.node_idx_arc.has_value())
+        arc_mid_nodes.insert(*edge.node_idx_arc);
+
+    for (const auto& edge : m_sketch.m_edges)
+    {
+      // Only consider non-arc edges (straight segments etc.).
+      if (edge.circle_arc)
+        continue;
+
+      if (!edge.node_idx_b.has_value())
+        continue;
+
+      const size_t a = edge.node_idx_a;
+      const size_t b = *edge.node_idx_b;
+
+      if (arc_mid_nodes.count(a) || arc_mid_nodes.count(b))
+        excluded_edges.insert(&edge);
+    }
+  }
+
+  bool changed = true;
+  while (changed)
+  {
+    changed = false;
+    std::unordered_set<const Sketch::Edge*> to_exclude;
+
+    // Find edges where at least one endpoint has degree 1 (considering only non-excluded edges)
+    for (const auto& edge : m_sketch.m_edges)
+    {
+      if (excluded_edges.count(&edge))
+        continue;
+
+      EZY_ASSERT(edge.node_idx_b.has_value());
+      size_t a = edge.node_idx_a;
+      size_t b = edge.node_idx_b.value();
+
+      // Count degree for each endpoint (excluding already excluded edges)
+      int degree_a = 0;
+      int degree_b = 0;
+      for (const auto& [neighbor, e] : adj_list[a])
+        if (!excluded_edges.count(e))
+          ++degree_a;
+
+      for (const auto& [neighbor, e] : adj_list[b])
+        if (!excluded_edges.count(e))
+          ++degree_b;
+
+      // If either endpoint has degree 1, this edge is dangling
+      if (degree_a == 1 || degree_b == 1)
+      {
+        to_exclude.insert(&edge);
+        changed = true;
+      }
+    }
+
+    excluded_edges.insert(to_exclude.begin(), to_exclude.end());
+  }
+  // Bridge edge removal logic (active). Excludes edges that purely connect separate cycles
+  // (both ends degree >=3 and no alternate path) from the adjacency used by the face walker.
+  // This is required for sketches like bridge.ezy that contain a "bridge" touching an inner
+  // triangular loop so that we obtain exactly two faces, one of which has the triangle as a
+  // hole. Dangling removal alone is not sufficient. The logic was previously disabled because
+  // an earlier version produced bad results on some nested rect holes; current state + other
+  // fixes (cw break, split rules, etc.) are being validated with the new regression test.
+  // Remove bridge edges that connect two separate cycles.
+  // A bridge edge connects two cycles and has both endpoints with degree >= 3.
+  // We detect this by checking if the edge is the only connection between two cycles.
+  std::unordered_set<const Sketch::Edge*> bridge_edges;
+  for (const auto& edge : m_sketch.m_edges)
+  {
+    if (excluded_edges.count(&edge))
+      continue;
+
+    EZY_ASSERT(edge.node_idx_b.has_value());
+    size_t a = edge.node_idx_a;
+    size_t b = edge.node_idx_b.value();
+
+    // Count degree for each endpoint (excluding already excluded edges)
+    int degree_a = 0;
+    int degree_b = 0;
+    for (const auto& [neighbor, e] : adj_list[a])
+      if (!excluded_edges.count(e))
+        ++degree_a;
+
+    for (const auto& [neighbor, e] : adj_list[b])
+      if (!excluded_edges.count(e))
+        ++degree_b;
+
+    // Bridge edges have both endpoints with degree >= 3
+    // They connect two separate cycles. We detect this by checking if removing the edge
+    // creates two disconnected components, each containing a cycle.
+    if (degree_a >= 3 && degree_b >= 3)
+    {
+      // Check if removing this edge disconnects the graph
+      // by seeing if we can reach 'b' from 'a' without using this edge
+      std::unordered_set<size_t> visited;
+      std::vector<size_t>        queue;
+      queue.push_back(a);
+      visited.insert(a);
+      bool can_reach_b = false;
+
+      for (size_t i = 0; i < queue.size() && !can_reach_b; ++i)
+      {
+        size_t curr = queue[i];
+        for (const auto& [neighbor, e] : adj_list[curr])
+        {
+          if (excluded_edges.count(e) || e == &edge)
+            continue;
+
+          if (neighbor == b)
+          {
+            can_reach_b = true;
+            break;
+          }
+
+          if (!visited.count(neighbor))
+          {
+            visited.insert(neighbor);
+            queue.push_back(neighbor);
+          }
+        }
+      }
+
+      // If we can't reach b from a without this edge, it's a bridge
+      // But we also need to verify both sides have cycles
+      if (!can_reach_b)
+      {
+        // Check if component containing 'a' has a cycle
+        auto has_cycle_in_component = [&](size_t start, const Sketch::Edge* exclude_edge) -> bool
+        {
+          std::unordered_set<size_t>          comp_visited;
+          std::function<bool(size_t, size_t)> dfs = [&](size_t curr, size_t parent) -> bool
+          {
+            comp_visited.insert(curr);
+            for (const auto& [neighbor, e] : adj_list[curr])
+            {
+              if (excluded_edges.count(e) || e == exclude_edge)
+                continue;
+
+              if (neighbor == parent)
+                continue;
+
+              if (comp_visited.count(neighbor))
+                // Found a back edge = cycle
+                return true;
+
+              if (dfs(neighbor, curr))
+                return true;
+            }
+            return false;
+          };
+
+          // Try starting from each neighbor
+          for (const auto& [neighbor, e] : adj_list[start])
+          {
+            if (excluded_edges.count(e) || e == exclude_edge)
+              continue;
+
+            comp_visited.clear();
+            comp_visited.insert(start);
+            if (dfs(neighbor, start))
+              return true;
+          }
+          return false;
+        };
+
+        bool a_has_cycle = has_cycle_in_component(a, &edge);
+        bool b_has_cycle = has_cycle_in_component(b, &edge);
+
+        // If both components have cycles, this edge is a bridge
+        if (a_has_cycle && b_has_cycle)
+          bridge_edges.insert(&edge);
+      }
+    }
+  }
+
+  // Add bridge edges to excluded set
+  excluded_edges.insert(bridge_edges.begin(), bridge_edges.end());
+
+  // Rebuild adjacency list excluding dangling and bridge edges
+  adj_list.clear();
+  for (const auto& edge : m_sketch.m_edges)
+  {
+    if (excluded_edges.count(&edge))
+      continue;
+
+    EZY_ASSERT(edge.node_idx_b.has_value());
+    size_t a = edge.node_idx_a;
+    size_t b = edge.node_idx_b.value();
+    adj_list[a].push_back({b, &edge});
+    adj_list[b].push_back({a, &edge});
+  }
+
+  std::unordered_set<std::pair<size_t, size_t>, Pair_hash> seen_edges;
+
+  const auto edge_outgoing_dir = [this](size_t idx_a, size_t idx_b) -> gp_Vec2d
+  {
+    gp_Vec2d ret(m_sketch.m_nodes[idx_a], m_sketch.m_nodes[idx_b]);
+    ret.Normalize();
+    return ret;
+  };
+
+  for (auto& [a_idx, edges] : adj_list)
+    for (auto& [b_idx, start_edge] : edges)
+    {
+      if (seen_edges.count(std::make_pair(a_idx, b_idx)))
+        continue;
+
+      size_t              start_idx = a_idx;
+      size_t              prev_idx  = a_idx;
+      size_t              curr_idx  = b_idx;
+      const Sketch::Edge* curr_edge = start_edge;
+      Face_edges          face;
+
+      for (;;)
+      {
+        face.push_back({*curr_edge, curr_edge->reversed(prev_idx, curr_idx)});
+
+        // Base case
+        if (curr_idx == start_idx)
+        {
+          EZY_ASSERT(face.size() > 2);
+
+          // Deliberately accept the cycle only if the left-most walker produced a traversal
+          // whose shoelace is positive under our is_face_clockwise_ convention. If not, break
+          // (skip pushing a face for this seed). This is not an error -- other seeds may still
+          // discover the same geometric cycle (or its sibling faces) in the matching orientation.
+          // The early break makes the set of accepted faces and their order in m_faces depend on
+          // m_edges list order (affected by appends and push_backs during splits) and on individual
+          // Edge node_idx_a/b values (set from pa/pb at creation time, i.e. draw direction for
+          // edges added via the GUI line tool). This dependency has been observed to be required
+          // for correct results in the GUI for cases such as "square then mid vertical splitter
+          // added lower-to-top vs. top-to-bottom" (the break version yields the expected rects
+          // on the sides the user sees; a force-to-cw normalization for every collected cycle
+          // re-introduces the direction asymmetry or malformed faces in that flow).
+          // The hole test (and bridge/dangling face tests) still pass because their edge addition
+          // order + walker seeds happen to produce matching orientations for both outer and inner
+          // boundaries.
+          if (!is_face_clockwise_(face))
+            break;
+
+          for (const Face_edge& e : face)
+          {
+            // Mark edge in both directions to avoid processing the same face twice
+            // when starting from different nodes
+            size_t start = e.start_nd_idx();
+            size_t end   = e.end_nd_idx();
+            seen_edges.insert(std::make_pair(start, end));
+            seen_edges.insert(std::make_pair(end, start));
+          }
+
+          auto f = create_face_shape_(face);
+          f->SetColor(Quantity_NOC_GRAY7);       // Base color
+          f->SetTransparency(0.5);               // 0.5 = 50% transparent (0.0 = opaque, 1.0 = invisible)
+          f->SetMaterial(Graphic3d_NOM_PLASTIC); // Optional: material for shading
+
+          m_sketch.m_ctx.Display(f, AIS_Shaded, AIS_Shape::SelectionMode(TopAbs_FACE), true);
+          m_sketch.m_ctx.Activate(f, AIS_Shape::SelectionMode(TopAbs_FACE));
+          // m_ctx.UpdateCurrentViewer();
+          m_faces.push_back(f);
+          break;
+        }
+
+        size_t              left_most_idx;
+        double              min_angle      = std::numeric_limits<double>::max();
+        const Sketch::Edge* left_most_edge = nullptr;
+        gp_Vec2d            edge_a         = edge_outgoing_dir(prev_idx, curr_idx);
+
+        for (auto& [next_idx, next_edge] : adj_list[curr_idx])
+        {
+          if (next_idx == prev_idx)
+            continue;
+
+          gp_Vec2d edge_b = edge_outgoing_dir(curr_idx, next_idx);
+          double   angle  = std::atan2(edge_b.Crossed(edge_a), edge_b.Dot(edge_a));
+          if (angle < min_angle)
+          {
+            min_angle      = angle;
+            left_most_idx  = next_idx;
+            left_most_edge = next_edge;
+          }
+        }
+        if (!left_most_edge)
+          // Invalid topology, probably a dangling edge
+          break;
+
+        prev_idx  = curr_idx;
+        curr_idx  = left_most_idx;
+        curr_edge = left_most_edge;
+      }
+    }
+
+  // Mark unused nodes as deleted so they're excluded from snapping operations.
+  // Nodes become unused when all edges that reference them are removed.
+  // Permanent add-node points are never auto-tombstoned when unused; user delete sets `deleted` and it stays.
+  for (size_t idx = 0, num = m_sketch.m_nodes.size(); idx < num; ++idx)
+    if (!m_sketch.m_nodes[idx].permanent)
+      m_sketch.m_nodes[idx].deleted = !used_nodes[idx];
+
+  // Book keeping
+  struct Face_meta
+  {
+    Sketch_face_shp_ptr           shp;
+    double                        area;
+    int                           parent_idx;
+    std::vector<const Face_meta*> holes;
+  };
+
+  std::vector<Face_meta> face_metas;
+  face_metas.reserve(m_faces.size());
+  for (Sketch_face_shp_ptr& face : m_faces)
+    face_metas.push_back({face, compute_face_area(face), -1});
+
+  // Sort faces by area (descending) to check larger faces first
+  std::sort(face_metas.begin(), face_metas.end(),
+            [](const Face_meta& a, const Face_meta& b)
+            {
+              return a.area > b.area; // Descending by area
+            });
+
+  // Classify faces
+  for (size_t face_i = 0, num = face_metas.size(); face_i < num; ++face_i)
+  {
+    Face_meta& face_a = face_metas[face_i];
+    for (size_t face_j = face_i + 1; face_j < num; ++face_j)
+    {
+      Face_meta& face_b = face_metas[face_j];
+      if (is_face_contained(face_b.shp->Shape(), face_a.shp->Shape()))
+        // Check if face_a is better (smaller) parent than the current one
+        if (face_b.parent_idx == -1 || face_a.area < face_metas[face_a.parent_idx].area)
+          face_b.parent_idx = static_cast<int>(face_i);
+    }
+  }
+
+  // Assign holes
+  for (const Face_meta& face : face_metas)
+    if (face.parent_idx != -1)
+      face_metas[face.parent_idx].holes.push_back(&face);
+
+  // Rebuild face shapes with their holes
+  for (Face_meta& face : face_metas)
+    if (face.holes.size())
+    {
+      EZY_ASSERT(face.shp->Shape().ShapeType() == TopAbs_FACE);
+      BRepBuilderAPI_MakeFace face_maker(TopoDS::Face(face.shp->Shape()));
+      for (const Face_meta* hole : face.holes)
+      {
+        EZY_ASSERT(hole->shp->Shape().ShapeType() == TopAbs_FACE);
+        TopoDS_Wire hole_wire = BRepTools::OuterWire(TopoDS::Face(hole->shp->Shape()));
+        // Winding order is important
+        hole_wire.Reverse();
+        face_maker.Add(hole_wire);
+      }
+      EZY_ASSERT(face_maker.IsDone());
+      // auto dbg_face = to_boost(face_maker.Face(), m_pln);
+      face.shp->SetShape(face_maker.Face());
+    }
+
+  // Use the nesting depth of the faces to define the face selection sensitivity
+  for (const Face_meta& face : face_metas)
+  {
+    int              nesting_depth = 0;
+    const Face_meta* curr_face     = &face;
+    for (; curr_face->parent_idx != -1; ++nesting_depth)
+      curr_face = &face_metas[curr_face->parent_idx];
+
+    m_sketch.m_ctx.SetSelectionSensitivity(face.shp, 0, nesting_depth + 1);
+  }
+
+  rebuild_dim_classifier_face_cache_();
+  m_sketch.purge_stale_length_dimensions_();
+  m_sketch.sync_permanent_node_annos_();
+  m_sketch.refresh_all_length_dimensions_();
+
+  if (m_sketch.is_current())
+    m_sketch.m_view.refresh_active_sketch_grid();
+}
+
+void Sketch_topo::rebuild_dim_classifier_face_cache_()
+{
+  m_dim_classifier_faces.clear();
+  m_dim_classifier_faces.reserve(m_faces.size());
+  for (const Sketch_face_shp_ptr& fp : m_faces)
+    m_dim_classifier_faces.push_back(TopoDS::Face(fp->Shape()));
+}
+
+size_t Sketch_topo::Face_edge::start_nd_idx() const
+{
+  if (!reversed)
+    return edge.node_idx_a;
+
+  EZY_ASSERT(edge.node_idx_b);
+  return *edge.node_idx_b;
+}
+
+size_t Sketch_topo::Face_edge::end_nd_idx() const
+{
+  if (reversed)
+    return edge.node_idx_a;
+
+  EZY_ASSERT(edge.node_idx_b);
+  return *edge.node_idx_b;
+}
+
+bool Sketch_topo::is_face_clockwise_(const Face_edges& face) const
+{
+  // Compute signed area using the shoelace formula
+  double signed_area = 0.0;
+  for (const Face_edge& e : face)
+  {
+    // `start_nd_idx` and `end_nd_idx` consider reversed edges
+    const gp_Pnt2d& p1 = m_sketch.m_nodes[e.start_nd_idx()];
+    const gp_Pnt2d& p2 = m_sketch.m_nodes[e.end_nd_idx()];
+    signed_area += (p1.X() * p2.Y()) - (p2.X() * p1.Y());
+  }
+  // signed_area *= 0.5;  // Divide by 2 for actual area
+
+  return signed_area > 0;
+}
+
+Sketch_face_shp_ptr Sketch_topo::create_face_shape_(const Face_edges& face)
+{
+  EZY_ASSERT(face.size() > 2);
+
+  // Create edges for the wire
+  BRepBuilderAPI_MakeWire wire_maker;
+
+  struct Circle_arc
+  {
+    std::optional<size_t> nd_idx_a;
+    std::optional<size_t> nd_idx_b;
+    std::optional<size_t> nd_idx_c;
+    std::optional<bool>   dbg_reversed;
+  };
+
+  std::map<Geom_TrimmedCurve*, Circle_arc> circle_arcs;
+  std::vector<gp_Pnt>                      node_verts;
+  std::optional<size_t>                    dbg_last_node_idx;
+
+  for (const Face_edge& e : face)
+  {
+    EZY_ASSERT(e.edge.node_idx_b);
+
+    auto add_node_vert_unique = [&](const gp_Pnt& pt)
+    {
+      if (node_verts.empty())
+        node_verts.push_back(pt);
+      else if (!node_verts.back().IsEqual(pt, Precision::Confusion()))
+        node_verts.push_back(pt);
+    };
+
+    if (e.edge.circle_arc.get())
+    {
+      Circle_arc& arc = circle_arcs[e.edge.circle_arc.get()];
+
+      auto try_arc_circle = [&](bool reversed)
+      {
+        if (!arc.dbg_reversed.has_value())
+          arc.dbg_reversed = reversed;
+        else
+          EZY_ASSERT(*arc.dbg_reversed == reversed);
+
+        if (arc.nd_idx_a && arc.nd_idx_b && arc.nd_idx_c)
+        {
+          gp_Pnt a = m_sketch.to_3d_(arc.nd_idx_a);
+          gp_Pnt b = m_sketch.to_3d_(arc.nd_idx_b);
+          gp_Pnt c = m_sketch.to_3d_(arc.nd_idx_c);
+
+          add_node_vert_unique(a);
+          add_node_vert_unique(b);
+          add_node_vert_unique(c);
+
+          Geom_TrimmedCurve_ptr arc_circle;
+          if (reversed)
+            arc_circle = GC_MakeArcOfCircle(c, b, a);
+          else
+            arc_circle = GC_MakeArcOfCircle(a, b, c);
+
+          BRepBuilderAPI_MakeEdge edge(arc_circle);
+          EZY_ASSERT(edge.IsDone());
+          wire_maker.Add(edge.Edge());
+        }
+      };
+
+      // Depending on the order of the node indexes in face,
+      // the first or second circle arc edge might come first.
+      if (e.edge.node_idx_arc)
+      {
+        // This is the first part of the circle arc.
+        EZY_ASSERT(e.edge.node_idx_arc);
+        EZY_ASSERT(*e.edge.node_idx_b == *e.edge.node_idx_arc);
+        arc.nd_idx_a = e.edge.node_idx_a;
+        arc.nd_idx_b = e.edge.node_idx_arc;
+        try_arc_circle(e.reversed);
+      }
+      else
+      {
+        // This is the second part of the circle arc.
+        arc.nd_idx_c = *e.edge.node_idx_b;
+        try_arc_circle(e.reversed);
+      }
+    }
+    else
+    {
+      gp_Pnt a = m_sketch.to_3d_(e.start_nd_idx());
+      gp_Pnt b = m_sketch.to_3d_(e.end_nd_idx());
+
+      add_node_vert_unique(a);
+      add_node_vert_unique(b);
+
+      BRepBuilderAPI_MakeEdge edge(a, b);
+      EZY_ASSERT(edge.IsDone());
+      wire_maker.Add(edge.Edge());
+    }
+  }
+
+  EZY_ASSERT(wire_maker.IsDone());
+  TopoDS_Wire wire = wire_maker.Wire();
+
+  // Create a face from the wire
+  BRepBuilderAPI_MakeFace face_maker(wire);
+  EZY_ASSERT(face_maker.IsDone());
+
+  // auto dbg_face = to_boost(face_maker.Face(), m_pln);
+
+  Sketch_face_shp* ret = new Sketch_face_shp(m_sketch, face_maker.Face());
+  std::swap(node_verts, ret->verts_3d);
+  return ret;
+}
